main/requirements.txt
Django==1.5.5
MySQL-python==1.2.5
South==0.8.4
argparse==1.2.1
distribute==0.6.34
django-ipware==0.0.8
django-lazysignup==0.12.2
django-social-auth==0.7.28
django-modeltranslation==0.8b1
httplib2==0.8
oauth2==1.5.211
prettytable==0.7.2
proso-geography-model==1.0.2
python-openid==2.2.5
wsgiref==0.1.2
